Fix socket timeouts which were causing jsonbot to drop/rejoin to irc
Not sure what the root cause was but two things seem to make a
difference:

- setting TCP keepalive on a socket

- changing socket timeout to infinity.

My understanding is that the main reason for this fix to work is
pushing socket timeout past the IRC PING/PONG rate. Otherwise
readline() on socket times out beore we get a PING/PONG and we rejoin.

We also set tcp keepalive to ensure the server is still there, even
though higher layer never times out by default.